Added tokens/authentication
Summary:
Right now, the only API call that is secured is /get_quotes -- for testing purposes. Will
secure the rest once we agree that this is how we want to do auth and that everything works
as intended.

application.py
--------------

GET /token ARGS: fbid, (fb)token
Calls validate (see below) to make sure that this is a valid fb user and he/she
has granted us permission. Creates a token (using tokenlib) that contains user_id (user.id not fbid)
and a random number from 1-100000 as well as a timeout (which I've set to 1 year, lul).

Returns a JSON object containing "user_id" and "access_token" fields.

================================
validate(method, user_id, token):
Passes the fb token to fb and checks if the fbid from fb matches the one that is passed.

The method arg is for the future if we want to do other sorts of validation, for now
there is just fb.

===============================
authorize_user(user_id, access_token):
This is the method that goes in each API call. Raises an AuthException if anything goes
wrong which gets returned as a JSON response. Otherwise, it will allow the call to continue.

model.py
--------

Just added a simple Access_Token model which contains "user_id" and "access_token" fields.

Test Plan: Push to server and have Rishi do the iOS part (for the sake of time).

Reviewers: mom, rishi

CC: jacob, juan

Maniphest Tasks: T63

Differential Revision: http://phabricator.echoapp.me/D99