use echoes as a proxy for quotes, only query echoes table. Fix chronological stuff
Summary:
Previously, we used to query the Quotes table and use Echoes as a proxy for the Quotes.echoers connection. However, this made it very difficult to get_quotes chronologically together with echoes.

To fix this, I changed the convention so that now we only query the Echoes table for both quotes and echoes. For each quote, we add a dummy echo so that we know where it stands chronologically with respect to the other echoes. There main difficulty in the whole process is that now we're passing Echo id's instead of Quote id's to the client, so we need to make sure to convert appropriately for all the quote-manipulation API calls.

Test Plan: testui

Reviewers: rishi, chris

CC: jacob, juan

Maniphest Tasks: T180

Differential Revision: http://phabricator.echoapp.me/D100